+++ src/main/webapp/components/gallery/gallery.js
@@ -3,7 +3,7 @@ import CardActionArea from '@material-ui/core/CardActionArea'
 import Dialog from '@material-ui/core/Dialog'
 import Typography from '@material-ui/core/Typography'
 import React from 'react'
-import Inspector from '../../inspector'
+import Inspector from '../inspector'
 import Thumbnail from '../../thumbnail/thumbnail'
 
 const GalleryItem = props => {
