+++ src/main/webapp/components/inspector/index.js
