+++ src/main/webapp/workspaces/visualizations.js
@@ -10,7 +10,7 @@ import { useSelectionInterface, useDrawInterface } from '../react-hooks'
 import { Layout, Provider, AddConfig, DragSource } from '../react-golden-layout'
 
 import Histogram from '../components/histogram/histogram'
-import Inspector from '../inspector/inspector'
+import Inspector from '../components/inspector/inspector'
 import ResultTable from '../results/results'
 import Gallery from '../components/gallery'
